presence-ai-backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
SpeechRecognition
pydantic
moviepy==1.0.3
google-generativeai
python-dotenv
